Update test_update_user and test_delete_user to use JWTauth token, Testy PUT i DELETE/users/{id} zaktualizowane po zabezpieczeniu endpointów - dodano auth_user fixtures oraz header Authorization

+++ Projekty/Task_Manager_API/tests/test_users.py
@@ -1,10 +1,7 @@
-# TODO: Napisz testy dla endpointow User
-
 
 
 
 def test_create_user(client):
-    """Test tworzenia uzytkownika."""
     
     test_user_data = {
         "username": "testuser",
@@ -27,9 +24,7 @@ def test_create_user(client):
 
 
 def test_get_users(client):
-    """Test pobierania listy uzytkownikow."""
-    # Najpierw utworz uzytkownika
-    # Potem pobierz liste
+    
     test_user_data = {
         "username": "testuser",
         "email": "test@example.com",
@@ -51,7 +46,7 @@ def test_get_users(client):
 
 
 def test_get_user_not_found(client):
-    """Test 404 dla nieistniejacego uzytkownika."""
+
     response = client.get("/users/999")
     assert response.status_code == 404
 
@@ -59,23 +54,16 @@ def test_get_user_not_found(client):
 
 
 
-def test_update_user(client):
-    """Test aktualizacji uzytkownika."""
+def test_update_user(client, auth_token, test_user):
 
-    test_user_data = {
-        "username": "testuser",
-        "email": "test@example.com",
-        "password": "examplePassword"
-
-    }
-
-    response = client.post("/auth/register", json=test_user_data)
-    assert response.status_code == 201
+    assert test_user["username"] == "test_data_username"
+    assert test_user["email"] == "example@exapmle.pl"
     
-    user_id = response.json()["id"]    
+    
+    user_id = test_user["id"]    
 
     update_data = {"username": "new_test_name"}
-    response = client.put(f"/users/{user_id}", json=update_data)
+    response = client.put(f"/users/{user_id}", json=update_data, headers={"Authorization": f"Bearer {auth_token}"})
     assert response.status_code == 200
     assert response.json()["username"] == "new_test_name"
 
@@ -84,21 +72,13 @@ def test_update_user(client):
 
 
 
-def test_delete_user(client):
-    """Test usuwania uzytkownika."""
+def test_delete_user(client, auth_token, test_user):
 
-    test_user_data = {
-        "username": "testuser",
-        "email": "test@example.com",
-        "password": "examplePassword"
+    assert test_user["username"] == "test_data_username"
+    assert test_user["email"] == "example@exapmle.pl"
 
-    }
-
-    response = client.post("/auth/register", json=test_user_data)
-    assert response.status_code == 201
-
-    user_id = response.json()["id"]
-    response = client.delete(f"/users/{user_id}")
+    user_id = test_user["id"]
+    response = client.delete(f"/users/{user_id}", headers={"Authorization": f"Bearer {auth_token}"})
     assert response.status_code == 204
 
     get_id = client.get(f"/users/{user_id}")
@@ -107,7 +87,7 @@ def test_delete_user(client):
 
 
 def test_create_duplicate_username(client):
-    """Test ze nie mozna utworzyc dwoch userow z tym samym username."""
+
     test_user_data = {
         "username": "testuser",
         "email": "test@example.com",
